1. projects/DeepResearchAgent/requirements.txt
python-dotenv>=1.0.1
rich>=13.9.4
litellm>=1.75.0
openai>=1.90.0
pillow>=10.4.0
numpy>=2.2.4
pynput>=1.8.1
transformers>=4.50.1
markdownify>=1.1.0
duckduckgo-search>=7.5.4
mammoth>=1.9.0
pandas>=2.2.3
pdfminer-six>=20250327
python-pptx>=1.0.2
puremagic>=1.28
pydub>=0.25.1
speechrecognition>=3.14.2
pathvalidate>=3.2.3
serpapi>=0.1.5
google-search-results>=2.4.2
datasets>=3.5.0
anthropic>=0.49.0
openpyxl>=3.1.5
baidusearch>=1.0.3
patchright>=1.51.3
pyautogui>=0.9.54
json5>=0.12.0
tenacity>=9.1.2
crawl4ai>=0.6.3
camelot-py>=1.0.0
toml>=0.10.2
xlrd>=0.7.1
googlesearch-python>=1.3.0
fastmcp>=2.8.1
httpx_aiohttp>=0.1.6
inflection>=0.5.1
mmengine>=0.10.7
markitdown>=0.1.3
firecrawl>=4.3.5
browser_use==0.1.48